backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==0.27.8
PyJWT==2.8.0
pydantic==2.4.2
httpx==0.25.0
